+++ app/src/main/java/com/example/foosball/GameActivity.java
@@ -93,18 +93,18 @@ public class GameActivity extends FullScreenActivity implements OnClickListener 
         return new Point (x,y);
     }
 
-    private Point getRandomPoint() {
-        Random r = new Random();
-        int minX = 0;
-        int maxX = findViewById(R.id.the_canvas).getWidth() - ((GameBoard)findViewById(R.id.the_canvas)).getSprite1Width();
-        int x = 0;
-        int minY = 0;
-        int maxY = findViewById(R.id.the_canvas).getHeight() - ((GameBoard)findViewById(R.id.the_canvas)).getSprite1Height();
-        int y = 0;
-        x = r.nextInt(maxX-minX+1)+minX;
-        y = r.nextInt(maxY-minY+1)+minY;
-        return new Point (x,y);
-    }
+//    private Point getRandomPoint() {
+//        Random r = new Random();
+//        int minX = 0;
+//        int maxX = findViewById(R.id.the_canvas).getWidth() - ((GameBoard)findViewById(R.id.the_canvas)).getSprite1Width();
+//        int x = 0;
+//        int minY = 0;
+//        int maxY = findViewById(R.id.the_canvas).getHeight() - ((GameBoard)findViewById(R.id.the_canvas)).getSprite1Height();
+//        int y = 0;
+//        x = r.nextInt(maxX-minX+1)+minX;
+//        y = r.nextInt(maxY-minY+1)+minY;
+//        return new Point (x,y);
+//    }
 
     synchronized public void initGfx() {
         ((GameBoard)findViewById(R.id.the_canvas)).resetStarField();
@@ -112,34 +112,34 @@ public class GameActivity extends FullScreenActivity implements OnClickListener 
                 pTeamBGoalie, pTeamBDefender1, pTeamBDefender2, pTeamBAttacker1, pTeamBAttacker2, pTeamBAttacker3;
 
         // Generate Positions
-        pTeamAGoalie = (findViewById(R.id.the_canvas).getWidth() * 0.05, findViewById(R.id.the_canvas).getHeight() * 0.50);
-        pTeamADefender1 = (findViewById(R.id.the_canvas).getWidth() * 0.20, findViewById(R.id.the_canvas).getHeight() * 0.625);
-        pTeamADefender2 = (findViewById(R.id.the_canvas).getWidth() * 0.20, findViewById(R.id.the_canvas).getHeight() * 0.375);
-        pTeamAAttacker1 = (findViewById(R.id.the_canvas).getWidth() * 0.60, findViewById(R.id.the_canvas).getHeight() * 0.75);
-        pTeamAAttacker2 = (findViewById(R.id.the_canvas).getWidth() * 0.60, findViewById(R.id.the_canvas).getHeight() * 0.50);
-        pTeamAAttacker3 = (findViewById(R.id.the_canvas).getWidth() * 0.60, findViewById(R.id.the_canvas).getHeight() * 0.25);
-        pTeamBGoalie = (findViewById(R.id.the_canvas).getWidth() * 0.95, findViewById(R.id.the_canvas).getHeight() * 0.50);
-        pTeamBDefender1 = (findViewById(R.id.the_canvas).getWidth() * 0.80, findViewById(R.id.the_canvas).getHeight() * 0.625);
-        pTeamADefender2 = (findViewById(R.id.the_canvas).getWidth() * 0.80, findViewById(R.id.the_canvas).getHeight() * 0.375);
-        pTeamBAttacker1 = (findViewById(R.id.the_canvas).getWidth() * 0.40, findViewById(R.id.the_canvas).getHeight() * 0.75);
-        pTeamBAttacker2 = (findViewById(R.id.the_canvas).getWidth() * 0.40, findViewById(R.id.the_canvas).getHeight() * 0.50);
-        pTeamBAttacker3 = (findViewById(R.id.the_canvas).getWidth() * 0.40, findViewById(R.id.the_canvas).getHeight() * 0.25);
-        pBall = (findViewById(R.id.the_canvas).getWidth() * 0.50, findViewById(R.id.the_canvas).getHeight() * 0.50);
+        pTeamAGoalie = new Point ((int) (findViewById(R.id.the_canvas).getWidth() * 0.5), (int) (findViewById(R.id.the_canvas).getHeight() * 0.5));
+//        pTeamADefender1 = (findViewById(R.id.the_canvas).getWidth() * 0.20, findViewById(R.id.the_canvas).getHeight() * 0.625);
+//        pTeamADefender2 = (findViewById(R.id.the_canvas).getWidth() * 0.20, findViewById(R.id.the_canvas).getHeight() * 0.375);
+//        pTeamAAttacker1 = (findViewById(R.id.the_canvas).getWidth() * 0.60, findViewById(R.id.the_canvas).getHeight() * 0.75);
+//        pTeamAAttacker2 = (findViewById(R.id.the_canvas).getWidth() * 0.60, findViewById(R.id.the_canvas).getHeight() * 0.50);
+//        pTeamAAttacker3 = (findViewById(R.id.the_canvas).getWidth() * 0.60, findViewById(R.id.the_canvas).getHeight() * 0.25);
+//        pTeamBGoalie = (findViewById(R.id.the_canvas).getWidth() * 0.95, findViewById(R.id.the_canvas).getHeight() * 0.50);
+//        pTeamBDefender1 = (findViewById(R.id.the_canvas).getWidth() * 0.80, findViewById(R.id.the_canvas).getHeight() * 0.625);
+//        pTeamBDefender2 = (findViewById(R.id.the_canvas).getWidth() * 0.80, findViewById(R.id.the_canvas).getHeight() * 0.375);
+//        pTeamBAttacker1 = (findViewById(R.id.the_canvas).getWidth() * 0.40, findViewById(R.id.the_canvas).getHeight() * 0.75);
+//        pTeamBAttacker2 = (findViewById(R.id.the_canvas).getWidth() * 0.40, findViewById(R.id.the_canvas).getHeight() * 0.50);
+//        pTeamBAttacker3 = (findViewById(R.id.the_canvas).getWidth() * 0.40, findViewById(R.id.the_canvas).getHeight() * 0.25);
+//        pBall = (findViewById(R.id.the_canvas).getWidth() * 0.50, findViewById(R.id.the_canvas).getHeight() * 0.50);
 
         // Set Positions
         ((GameBoard)findViewById(R.id.the_canvas)).TeamAGoalie.setPoint(pTeamAGoalie.x, pTeamAGoalie.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamADefender1.setPoint(pTeamADefender1.x, pTeamADefender1.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamADefender2.setPoint(pTeamADefender2.x, pTeamADefender2.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker1.setPoint(pTeamAAttacker1.x, pTeamAAttacker1.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker2.setPoint(pTeamAAttacker2.x, pTeamAAttacker2.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker3.setPoint(pTeamAAttacker3.x, pTeamAAttacker3.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamBGoalie.setPoint(pTeamBGoalie.x, pTeamBGoalie.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamBDefender1.setPoint(pTeamBDefender1.x, pTeamBDefender1.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamBDefender2.setPoint(pTeamBDefender2.x, pTeamBDefender2.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker1.setPoint(pTeamBAttacker1.x, pTeamBAttacker1.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker2.setPoint(pTeamBAttacker2.x, pTeamBAttacker2.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker3.setPoint(pTeamBAttacker3.x, pTeamBAttacker3.y);
-        ((GameBoard)findViewById(R.id.the_canvas)).setBall(pBall.x, pBall.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamADefender1.setPoint(TeamADefender1.x, TeamADefender1.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamADefender2.setPoint(TeamADefender2.x, TeamADefender2.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker1.setPoint(TeamAAttacker1.x, TeamAAttacker1.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker2.setPoint(TeamAAttacker2.x, TeamAAttacker2.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker3.setPoint(TeamAAttacker3.x, TeamAAttacker3.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamBGoalie.setPoint(TeamBGoalie.x, TeamBGoalie.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamBDefender1.setPoint(TeamBDefender1.x, TeamBDefender1.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamBDefender2.setPoint(TeamBDefender2.x, TeamBDefender2.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker1.setPoint(TeamBAttacker1.x, TeamBAttacker1.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker2.setPoint(TeamBAttacker2.x, TeamBAttacker2.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker3.setPoint(TeamBAttacker3.x, TeamBAttacker3.y);
+//        ((GameBoard)findViewById(R.id.the_canvas)).setBall(pBall.x, pBall.y);
 
 //        sprite1Velocity = getRandomVelocity();
 //        sprite2Velocity = new Point(1,1);
@@ -174,6 +174,7 @@ public class GameActivity extends FullScreenActivity implements OnClickListener 
 //
 //            //Add our call to increase or decrease velocity
 //            updateVelocity();
+
 //            Point sprite1 = new Point (((GameBoard)findViewById(R.id.the_canvas)).getSprite1X(),
 //                    ((GameBoard)findViewById(R.id.the_canvas)).getSprite1Y()) ;
 //            Point sprite2 = new Point (((GameBoard)findViewById(R.id.the_canvas)).getSprite2X(),
@@ -207,19 +208,19 @@ public class GameActivity extends FullScreenActivity implements OnClickListener 
 //                ballVelocity.y *= -1;
 //            }
 
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamAGoalie.setPoint(pTeamAGoalie.x, pTeamAGoalie.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamADefender1.setPoint(pTeamADefender1.x, pTeamADefender1.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamADefender2.setPoint(pTeamADefender2.x, pTeamADefender2.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker1.setPoint(pTeamAAttacker1.x, pTeamAAttacker1.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker2.setPoint(pTeamAAttacker2.x, pTeamAAttacker2.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker3.setPoint(pTeamAAttacker3.x, pTeamAAttacker3.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamBGoalie.setPoint(pTeamBGoalie.x, pTeamBGoalie.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamBDefender1.setPoint(pTeamBDefender1.x, pTeamBDefender1.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamBDefender2.setPoint(pTeamBDefender2.x, pTeamBDefender2.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker1.setPoint(pTeamBAttacker1.x, pTeamBAttacker1.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker2.setPoint(pTeamBAttacker2.x, pTeamBAttacker2.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker3.setPoint(pTeamBAttacker3.x, pTeamBAttacker3.y);
-            ((GameBoard)findViewById(R.id.the_canvas)).setBall(ball.x, ball.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamAGoalie.setPoint(TeamAGoalie.x, TeamAGoalie.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamADefender1.setPoint(pTeamADefender1.x, pTeamADefender1.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamADefender2.setPoint(pTeamADefender2.x, pTeamADefender2.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker1.setPoint(pTeamAAttacker1.x, pTeamAAttacker1.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker2.setPoint(pTeamAAttacker2.x, pTeamAAttacker2.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamAAttacker3.setPoint(pTeamAAttacker3.x, pTeamAAttacker3.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamBGoalie.setPoint(pTeamBGoalie.x, pTeamBGoalie.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamBDefender1.setPoint(pTeamBDefender1.x, pTeamBDefender1.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamBDefender2.setPoint(pTeamBDefender2.x, pTeamBDefender2.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker1.setPoint(pTeamBAttacker1.x, pTeamBAttacker1.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker2.setPoint(pTeamBAttacker2.x, pTeamBAttacker2.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).TeamBAttacker3.setPoint(pTeamBAttacker3.x, pTeamBAttacker3.y);
+//            ((GameBoard)findViewById(R.id.the_canvas)).setBall(ball.x, ball.y);
             ((GameBoard)findViewById(R.id.the_canvas)).invalidate();
             frame.postDelayed(frameUpdate, FRAME_RATE);
         }